+++ src/consts/urls.js
@@ -51,8 +51,7 @@ export const IBC_BALANCES_DENOM_URL = `${BASE_URL}/ibc/balances/denom`;
 export const IBC_CHANNEL_STATUS_URL = `${BASE_URL}/ibc/channels/status`;
 export const IBC_DENOMS_ALL_URL = `${BASE_URL}/ibc/denoms/all`;
 // -- Order book
-export const ORDER_BOOK_URL =
-  'https://www.dlob.io/aggregator/external/api/v1/order-books/pb18vd8fpwxzck93qlwghaj6arh4p7c5n894vnu5g';
+export const ORDER_BOOK_URL = 'https://www.dlob.io/gecko/external/api/v1/exchange';
 // -- NFT
 export const NFT_URL = `${BASE_URL}/nft/scope`;
 // -- Social Media/Contact Information