+++ src/redux/actions/orderbookActions.js
@@ -9,9 +9,16 @@ export const GET_PRICE_HISTORY = 'ORDERBOOK::GET_PRICE_HISTORY';
 // Actions
 // - API Calls
 export const getDailyPrice = () => async dispatch =>
-  ajaxGet(GET_DAILY_PRICE, dispatch, `${ORDER_BOOK_URL}/daily-price`);
+  ajaxGet(GET_DAILY_PRICE, dispatch, `${ORDER_BOOK_URL}/tickers`);
 
+// parameter format is dd-mm-yyyy
 export const getPriceHistory =
-  (period = 'DAY') =>
+  (start_time = '', end_time = '') =>
   async dispatch =>
-    ajaxGet(GET_PRICE_HISTORY, dispatch, `${ORDER_BOOK_URL}/price-history?period=${period}`);
+    ajaxGet(
+      GET_PRICE_HISTORY,
+      dispatch,
+      `${ORDER_BOOK_URL}/historical_trades?ticker_id=HASH_USD&type=buy${
+        start_time && `&start_time=${start_time}`
+      }${end_time && `&end_time=${end_time}`}`
+    );